Build Graph method moved
